refactor(Results): add Diagram
